backend/requirements.txt
Django==5.1.3
psycopg>=3.1.8
Pillow

djangorestframework
markdown
django-filter
django-cors-headers
django-environ

stripe

requests